feat(ui): give each tasks dialog its own project scope (#326)
The board and quick-add each seed a scope on their own open and keep it for that
open, so a background session switch no longer retargets what the user is
reading. Neither is a dead click any more: with no project resolved the dialog
offers the project list where the board or the form would be, and a project
deleted while open falls back to the same list.

The quick-add form moves to QuickTaskForm so the dialog can hold both states
inside one DialogContent — swapping a Dialog root as it closes is what strands
pointer-events on the body.
