+++ packages/ui/src/features/tasks/TasksModalHost.tsx
@@ -2,17 +2,22 @@
  * TasksModalHost — single app-root host for the Tasks full-view modal and
  * the QuickTaskDialog. Driven by useTasksModal (zustand store).
  *
- * Resolves projectId from useActiveIdentity(). Registers ⌘⇧T → openQuick().
- * Listens for `mf:open-tasks` custom event (dispatched by SidebarHeader TasksBtn).
- * Loads on mount (so the first open has data) and refetches on the
- * open/quick-add rising edge (so externally-made changes are reflected — the
- * todos store has no WS event).
- * Mounted once in AppShell's outlet block.
+ * Each dialog holds its own per-open project scope, seeded from the sidebar
+ * filter on the rising edge of its own open. The two never share a pick, and
+ * neither follows a background session switch. Opening always produces a
+ * surface: with no project resolved the dialog shows the project list instead
+ * of the board, so the sidebar entry and ⌘⇧T can no longer be dead clicks.
+ *
+ * Registers ⌘⇧T → openQuick(). Listens for `mf:open-tasks` (dispatched by
+ * SidebarHeader TasksBtn). Mounted once in AppShell's outlet block.
  */
-import React, { useEffect, useRef } from 'react';
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog';
+import React, { useEffect } from 'react';
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog';
 import { cn } from '@/lib/utils';
-import { useActiveIdentity } from '@/features/sessions/use-active-identity';
+import { useSessionFilters } from '@/store/session-filters';
+import { useProjects } from '@/features/sessions/use-projects';
+import { useModalProjectScope } from '@/features/project-scope/use-modal-project-scope';
+import { ProjectPickList } from '@/features/project-scope/ProjectPickList';
 import { useTasksModal } from './use-tasks-modal';
 import { useStartTodoSession } from './use-start-todo-session';
 import { useTodosStore } from './use-todos-store';
@@ -23,31 +28,23 @@ interface Props {
   port: number;
 }
 
-export function TasksModalHost({ port }: Props): React.ReactElement | null {
+export function TasksModalHost({ port }: Props): React.ReactElement {
   const { open, quickOpen, closeModal, openModal, openQuick, closeQuick } = useTasksModal();
-  const { projectId } = useActiveIdentity();
-  const startSession = useStartTodoSession(port, projectId);
-  const { load, view } = useTodosStore();
-  const prevQuick = useRef(false);
+  const { projects } = useProjects();
+  const filterProjectId = useSessionFilters((s) => s.filterProjectId);
+  const board = useModalProjectScope(open);
+  const quick = useModalProjectScope(quickOpen);
+  const view = useTodosStore((s) => s.view);
 
-  // Eagerly load todos so the modal and quick-add show correct data the
-  // first time either one opens.
-  useEffect(() => {
-    if (!projectId || !port) return;
-    void load(port, projectId);
-  }, [port, projectId, load]);
+  // A project deleted while a modal is open leaves its scope pointing at
+  // nothing; falling back to the pick list keeps the surface honest without an
+  // effect racing the render.
+  const known = (id: string | null): string | null =>
+    id !== null && projects.some((project) => project.id === id) ? id : null;
+  const boardProjectId = known(board.projectId);
+  const quickProjectId = known(quick.projectId);
 
-  // Refetch on the quick-add rising edge. The store has no WS event
-  // (single-window refetch-on-mutation), so a change made outside this window
-  // — agent sessions, another window, direct DB writes — would otherwise show
-  // boot-time statuses. The full modal needs no edge here: TasksBoard mounts
-  // fresh on each open (Radix unmounts DialogContent) and loads itself.
-  useEffect(() => {
-    if (!projectId || !port) return;
-    const justOpened = quickOpen && !prevQuick.current;
-    prevQuick.current = quickOpen;
-    if (justOpened) void load(port, projectId);
-  }, [quickOpen, projectId, port, load]);
+  const startSession = useStartTodoSession(port, boardProjectId ?? undefined);
 
   // ⌘⇧T → open quick-add dialog
   useEffect(() => {
@@ -70,8 +67,6 @@ export function TasksModalHost({ port }: Props): React.ReactElement | null {
     return () => window.removeEventListener('mf:open-tasks', handleOpenTasks);
   }, [openModal]);
 
-  if (!projectId) return null;
-
   return (
     <>
       {/* Full-view Tasks modal */}
@@ -82,29 +77,56 @@ export function TasksModalHost({ port }: Props): React.ReactElement | null {
         }}
       >
         <DialogContent
-          showCloseButton={false}
+          showCloseButton={boardProjectId === null}
           className={cn(
-            'flex max-h-[85vh] min-h-[480px] w-full flex-col gap-0 p-0 transition-[width] duration-[180ms] ease-out',
-            view === 'list' ? 'sm:max-w-[880px]' : 'w-[90vw] sm:max-w-[1200px]',
+            boardProjectId === null
+              ? 'sm:max-w-sm'
+              : 'flex max-h-[85vh] min-h-[480px] w-full flex-col gap-0 p-0 transition-[width] duration-[180ms] ease-out',
+            boardProjectId !== null && (view === 'list' ? 'sm:max-w-[880px]' : 'w-[90vw] sm:max-w-[1200px]'),
           )}
         >
-          <DialogHeader className="sr-only">
-            <DialogTitle>Tasks</DialogTitle>
-          </DialogHeader>
-          <TasksBoard
-            port={port}
-            projectId={projectId}
-            onClose={closeModal}
-            onStartSession={(todo) => {
-              closeModal();
-              void startSession(todo.id, todo.status);
-            }}
-          />
+          {boardProjectId === null ? (
+            <>
+              <DialogHeader>
+                <DialogTitle>Tasks</DialogTitle>
+                <DialogDescription>Pick the project whose board you want to open.</DialogDescription>
+              </DialogHeader>
+              <ProjectPickList
+                surface="tasks-board"
+                projects={projects}
+                filterProjectId={filterProjectId}
+                onSelect={board.setProjectId}
+              />
+            </>
+          ) : (
+            <>
+              <DialogHeader className="sr-only">
+                <DialogTitle>Tasks</DialogTitle>
+              </DialogHeader>
+              <TasksBoard
+                port={port}
+                projectId={boardProjectId}
+                onClose={closeModal}
+                onStartSession={(todo) => {
+                  closeModal();
+                  void startSession(todo.id, todo.status);
+                }}
+              />
+            </>
+          )}
         </DialogContent>
       </Dialog>
 
       {/* Quick-add dialog */}
-      <QuickTaskDialog port={port} projectId={projectId} open={quickOpen} onClose={closeQuick} />
+      <QuickTaskDialog
+        port={port}
+        projectId={quickProjectId}
+        projects={projects}
+        filterProjectId={filterProjectId}
+        onSelectProject={quick.setProjectId}
+        open={quickOpen}
+        onClose={closeQuick}
+      />
     </>
   );
 }
